requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==46.0.2
daphne==4.2.1
distro==1.9.0
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
google-auth==2.41.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.1
openai==2.2.0
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.1.1
PyYAML==6.0.3
redis==6.4.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
zope.interface==8.0.1
